// src/app/models/project/project_policy.ts
export class ProjectPolicyModel {
    constructor() {
     
    }
       
    ppol_code: string = "";
    ppol_name_th: string = "";
    ppol_name_en: string = "";
    ppol_ot: string = "";
    ppol_allw: string = "";
    ppol_dg: string = "";
    ppol_leave: string = "";
    ppol_late: string = "";

    modified_by: string = "";
    modified_date: Date = new Date();
    index: number = 0;
    select: boolean = false;    
   
  }
  