// src/app/models/project/cost.ts
export class CostModel {
    constructor() {
             
    }
    
    cost_id: string = "";
    cost_code: string = "";
    cost_name_th: string = "";
    cost_name_en: string = "";
    cost_type: string = "";
    cost_allwcode: string = "";

    cost_auto: boolean = false;

    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;
    select: boolean = false;    
  }
  