// src/app/models/project/project_position.ts
export class PrjectPositionModel {
    constructor() {
             
    }
       
    pos_code: string = "";
    pos_name_th: string = "";
    pos_name_en: string = "";

    pos_type: string = "";

    pos_emp: number = 0;

    shift_code: string = "";
    shift_n: string = "";
    shift_working: number = 26;
    shift_hrs: number = 8;
    shift_overtime: number = 0;
    pol_allw_code: string = "";
    pol_slip_code: string = "";
    pol_uniform_code: string = "";
      
    allow1: number = 0;
    allow2: number = 0;
    allow4: number = 0;
    allow5: number = 0;
    allow6: number = 0;

    allow_emp: number = 0;
    allow_total: number = 0;

    ot_auto: boolean = false;   

    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;
    select: boolean = false;    
   

  }
  