// src/app/models/project/project_jobtype.ts
export class PrjectJobtypeModel {
    constructor() {
             
    }
       
    pjobtype_code: string = "";
    pjobtype_name_th: string = "";
    pjobtype_name_en: string = "";
   
    modified_by: string = "";
    modified_date: Date = new Date();
    index: number = 0;
    select: boolean = false;    
   

  }
  